refactor: вынести валидацию данных в пакет validator
