+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -1,25 +1,23 @@
 package ru.yandex.practicum.filmorate.storage.film;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import lombok.extern.slf4j.Slf4j;
+import ru.yandex.practicum.filmorate.validator.FilmValidation;
 
-
-import java.time.LocalDate;
-import java.time.Month;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
 @Component
 @Slf4j
+@RequiredArgsConstructor
 public class InMemoryFilmStorage implements FilmStorage {
-    private static final int MAX_DESCRIPTION_LENGTH = 200;
-    private static final LocalDate BIRTHDAY_MOVIE = LocalDate.of(1895, Month.DECEMBER, 28);
-
+    private final FilmValidation validation;
     private int counterId;
     private final Map<Integer, Film> films = new HashMap<>();
 
@@ -32,7 +30,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     public ResponseEntity<Film> createFilm(Film film) {
         log.info("Получен запрос к эндпойнту: 'POST /films'");
         try {
-            isValid(film);
+            validation.isValid(film);
         } catch (ValidationException e) {
             log.warn("Фильм не прошёл валидацию по причине: " + e.getMessage());
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(film);
@@ -57,16 +55,5 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     private void getNextId() {
         counterId++;
-    }
-
-    public void isValid(Film film) throws ValidationException {
-        if (film.getName() == null || film.getName().isBlank())
-            throw new ValidationException("название не может быть пустым.");
-        if (film.getDescription().length() > MAX_DESCRIPTION_LENGTH)
-            throw new ValidationException("максимальная длина описания — 200 символов");
-        if (film.getReleaseDate().isBefore(BIRTHDAY_MOVIE))
-            throw new ValidationException("дата релиза — не раньше 28 декабря 1895 года");
-        if (film.getDuration() <= 0)
-            throw new ValidationException("продолжительность фильма должна быть положительной");
     }
 }
