+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -1,21 +1,23 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.validator.UserValidation;
 
-import java.time.LocalDate;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
 @Slf4j
 @Component
+@RequiredArgsConstructor
 public class InMemoryUserStorage implements UserStorage {
-
+    private final UserValidation validation;
     private int counterId;
     private final Map<Integer, User> users = new HashMap<>();
 
@@ -28,7 +30,7 @@ public class InMemoryUserStorage implements UserStorage {
     public ResponseEntity<User> createUser(User user) {
         log.info("Получен запрос к эндпойнту: 'POST /users'");
         try {
-            isValid(user);
+            validation.isValid(user);
         } catch (ValidationException e) {
             log.warn("Пользователь не прошёл валидацию по причине: " + e.getMessage());
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(user);
@@ -53,15 +55,5 @@ public class InMemoryUserStorage implements UserStorage {
 
     private void getNextId() {
         counterId++;
-    }
-
-    public void isValid(User user) throws ValidationException {
-        if (user.getEmail() == null || user.getEmail().isBlank() || !user.getEmail().contains("@"))
-            throw new ValidationException("электронная почта не может быть пустой и должна содержать символ @");
-        if (user.getLogin() == null || user.getLogin().isBlank() || user.getLogin().split(" ").length > 1)
-            throw new ValidationException("логин не может быть пустым и содержать пробелы");
-        if (user.getName() == null || user.getName().isBlank()) user.setName(user.getLogin());
-        if (user.getBirthday().isAfter(LocalDate.now()))
-            throw new ValidationException("дата рождения не может быть в будущем");
     }
 }
