+++ src/main/java/ru/yandex/practicum/filmorate/validator/UserValidation.java
@@ -1,0 +1,20 @@
+package ru.yandex.practicum.filmorate.validator;
+
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.time.LocalDate;
+
+@Component
+public class UserValidation {
+    public void isValid(User user) throws ValidationException {
+        if (user.getEmail() == null || user.getEmail().isBlank() || !user.getEmail().contains("@"))
+            throw new ValidationException("электронная почта не может быть пустой и должна содержать символ @");
+        if (user.getLogin() == null || user.getLogin().isBlank() || user.getLogin().split(" ").length > 1)
+            throw new ValidationException("логин не может быть пустым и содержать пробелы");
+        if (user.getName() == null || user.getName().isBlank()) user.setName(user.getLogin());
+        if (user.getBirthday().isAfter(LocalDate.now()))
+            throw new ValidationException("дата рождения не может быть в будущем");
+    }
+}
